# docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.ValidateRequest:
    properties:
      json:
        example: '{"key":"value"}'
        type: string
      type:
        example: generic
        type: string
    type: object
host: localhost:3001
info:
  contact:
    name: Support
    url: https://fixfx.wiki/discord
  description: Backend API for FixFX documentation and tools
  title: FixFX API
  version: 0.1.0
paths:
  /artifacts/changes:
    get:
      description: Get changelog/commits between two artifact versions
      parameters:
      - description: Base version for comparison
        in: query
        name: base
        type: string
      - description: Head version for comparison
        in: query
        name: head
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get artifact changelog
      tags:
      - Artifacts
  /artifacts/check:
    get:
      description: Check the status and availability of an artifact
      parameters:
      - description: Version to check
        in: query
        name: version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Check artifact status
      tags:
      - Artifacts
  /artifacts/fetch:
    get:
      description: Fetch FiveM/RedM server artifacts with filtering and pagination
      parameters:
      - default: all
        description: Platform (windows/linux/all)
        in: query
        name: platform
        type: string
      - description: Specific version to fetch
        in: query
        name: version
        type: string
      - description: Filter by status (recommended/latest/active/deprecated/eol)
        in: query
        name: status
        type: string
      - default: version
        description: Sort by (version/date)
        in: query
        name: sortBy
        type: string
      - default: desc
        description: Sort order (asc/desc)
        in: query
        name: sortOrder
        type: string
      - default: 50
        description: Results per page
        in: query
        name: limit
        type: integer
      - default: 0
        description: Pagination offset
        in: query
        name: offset
        type: integer
      - default: true
        description: Include end-of-life artifacts
        in: query
        name: includeEol
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get artifacts
      tags:
      - Artifacts
  /artifacts/version/{version}:
    get:
      description: Fetch a specific artifact version
      parameters:
      - description: Artifact version
        in: path
        name: version
        required: true
        type: string
      - description: Platform filter
        in: query
        name: platform
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get artifact by version
      tags:
      - Artifacts
...
  /validator/validate:
    post:
      consumes:
      - application/json
      description: Validate JSON with optional txAdmin embed/config schema validation
      parameters:
      - description: JSON to validate
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.ValidateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Validate JSON
      tags:
      - Validator
schemes:
- http
- https
swagger: "2.0"